split code to generate tensor specs into separate file
